# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThreatLens API",
    "summary": "ThreatLens API contract.",
    "description": "The published API contract is versioned under `/v1` on the backend service and `/api/v1` through the web proxy. Legacy unversioned routes remain available for compatibility but are intentionally excluded from the OpenAPI schema.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Backend service root"
    },
    {
      "url": "/api",
      "description": "Web reverse proxy root"
    }
  ],
  "paths": {
...
    "/v1/stats/overview": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Stats Overview",
        "operationId": "get_stats_overview_v1_stats_overview_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "feed_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/feed-timeseries": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Feed Timeseries",
        "operationId": "get_feed_timeseries_v1_stats_feed_timeseries_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "feed_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Ids"
            }
          },
          {
            "name": "top_feeds",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Top Feeds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedTimeSeriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/activity-heatmap": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Activity Heatmap",
        "operationId": "get_activity_heatmap_v1_stats_activity_heatmap_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "feed_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityHeatmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/signal-radar": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Signal Radar",
        "operationId": "get_signal_radar_v1_stats_signal_radar_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "feed_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalRadarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityHeatmapResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "window_start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start At"
          },
          "window_end_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window End At"
          },
          "bucket_unit": {
            "type": "string",
            "enum": [
              "hour",
              "day"
            ],
            "title": "Bucket Unit"
          },
          "bucket_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bucket Labels"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ActivityHeatmapDayRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "max_count": {
            "type": "integer",
            "title": "Max Count"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "window_days",
          "window_start_at",
          "window_end_at",
          "bucket_unit",
          "bucket_labels",
          "rows",
          "max_count"
        ],
        "title": "ActivityHeatmapResponse"
      },
      "ActivitySummary": {
        "properties": {
          "items_last_24h": {
            "type": "integer",
            "title": "Items Last 24H"
          },
          "items_last_7d": {
            "type": "integer",
            "title": "Items Last 7D"
          },
          "items_last_30d": {
            "type": "integer",
            "title": "Items Last 30D"
          }
        },
        "type": "object",
        "required": [
          "items_last_24h",
          "items_last_7d",
          "items_last_30d"
        ],
        "title": "ActivitySummary"
      },
...
      "DailyVolumePoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "count"
        ],
        "title": "DailyVolumePoint"
      },
      "DerivedSummary": {
        "properties": {
          "extraction_success_rate_pct": {
            "type": "number",
            "title": "Extraction Success Rate Pct"
          },
          "error_rate_pct": {
            "type": "number",
            "title": "Error Rate Pct"
          },
          "avg_items_per_day_window": {
            "type": "number",
            "title": "Avg Items Per Day Window"
          }
        },
        "type": "object",
        "required": [
          "extraction_success_rate_pct",
          "error_rate_pct",
          "avg_items_per_day_window"
        ],
        "title": "DerivedSummary"
      },
...
      "FeedTimeSeriesResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "window_start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start At"
          },
          "window_end_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window End At"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/FeedTimeSeriesSeries"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "window_days",
          "window_start_at",
          "window_end_at",
          "series"
        ],
        "title": "FeedTimeSeriesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignalRadarResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "window_start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start At"
          },
          "window_end_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window End At"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "max_count": {
            "type": "integer",
            "title": "Max Count"
          },
          "axes": {
            "items": {
              "$ref": "#/components/schemas/SignalRadarAxisPoint"
            },
            "type": "array",
            "title": "Axes"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "window_days",
          "window_start_at",
          "window_end_at",
          "total",
          "max_count",
          "axes"
        ],
        "title": "SignalRadarResponse"
      },
...
      "StatsOverviewResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "window_start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start At"
          },
          "window_end_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window End At"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsSummary"
          },
          "activity": {
            "$ref": "#/components/schemas/ActivitySummary"
          },
          "derived": {
            "$ref": "#/components/schemas/DerivedSummary"
          },
          "status_breakdown": {
            "items": {
              "$ref": "#/components/schemas/StatusPoint"
            },
            "type": "array",
            "title": "Status Breakdown"
          },
          "daily_volume": {
            "items": {
              "$ref": "#/components/schemas/DailyVolumePoint"
            },
            "type": "array",
            "title": "Daily Volume"
          },
          "feed_breakdown": {
            "items": {
              "$ref": "#/components/schemas/FeedStats"
            },
            "type": "array",
            "title": "Feed Breakdown"
          },
          "top_domains": {
            "items": {
              "$ref": "#/components/schemas/DomainPoint"
            },
            "type": "array",
            "title": "Top Domains"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "window_days",
          "window_start_at",
          "window_end_at",
          "totals",
          "activity",
          "derived",
          "status_breakdown",
          "daily_volume",
          "feed_breakdown",
          "top_domains"
        ],
        "title": "StatsOverviewResponse"
      },
      "StatusPoint": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "status",
          "count"
        ],
        "title": "StatusPoint"
      },
...
      "TotalsSummary": {
        "properties": {
          "feeds_total": {
            "type": "integer",
            "title": "Feeds Total"
          },
          "feeds_enabled": {
            "type": "integer",
            "title": "Feeds Enabled"
          },
          "feeds_disabled": {
            "type": "integer",
            "title": "Feeds Disabled"
          },
          "items_total": {
            "type": "integer",
            "title": "Items Total"
          },
          "items_new": {
            "type": "integer",
            "title": "Items New"
          },
          "items_content_fetched": {
            "type": "integer",
            "title": "Items Content Fetched"
          },
          "items_error": {
            "type": "integer",
            "title": "Items Error"
          },
          "articles_total": {
            "type": "integer",
            "title": "Articles Total"
          }
        },
        "type": "object",
        "required": [
          "feeds_total",
          "feeds_enabled",
          "feeds_disabled",
          "items_total",
          "items_new",
          "items_content_fetched",
          "items_error",
          "articles_total"
        ],
        "title": "TotalsSummary"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}